[lib] Refactor json-schema libs
