--- packages/form/src/lib/json-schema-compare/index.ts
@@ -1,1 +1,0 @@
-export * from "./json-schema-compare.js";
