--- packages/form/src/lib/json-schema-merge/index.ts
@@ -1,2 +1,0 @@
-export * from "./json-schema-merge.js";
-export * from "./all-of-merge.js";
