+++ packages/form/src/lib/json-schema/merge/index.ts
@@ -1,0 +1,2 @@
+export * from "./merge.js";
+export * from "./all-of-merge.js";
