+++ packages/form/src/lib/traverser.ts
@@ -1,0 +1,4 @@
+export interface Visitor<Node, Context, R> {
+  onEnter?: (node: Node, ctx: Context) => Generator<R>;
+  onLeave?: (node: Node, ctx: Context) => Generator<R>;
+}
